+++ lib/view_component/base.rb
@@ -10,6 +10,7 @@ require "view_component/errors"
 require "view_component/inline_template"
 require "view_component/preview"
 require "view_component/slotable"
+require "view_component/slotable_default"
 require "view_component/translatable"
 require "view_component/with_content_helper"
 require "view_component/use_helpers"
@@ -28,6 +29,7 @@ module ViewComponent
     end
 
     include ViewComponent::InlineTemplate
+    include ViewComponent::UseHelpers
     include ViewComponent::Slotable
     include ViewComponent::Translatable
     include ViewComponent::WithContentHelper
@@ -104,7 +106,14 @@ module ViewComponent
       before_render
 
       if render?
-        render_template_for(@__vc_variant).to_s + output_postamble
+        # Avoid allocating new string when output_preamble and output_postamble are blank
+        rendered_template = safe_render_template_for(@__vc_variant).to_s
+
+        if output_preamble.blank? && output_postamble.blank?
+          rendered_template
+        else
+          safe_output_preamble + rendered_template + safe_output_postamble
+        end
       else
         ""
       end
@@ -151,11 +160,18 @@ module ViewComponent
       end
     end
 
+    # Optional content to be returned before the rendered template.
+    #
+    # @return [String]
+    def output_preamble
+      @@default_output_preamble ||= "".html_safe
+    end
+
     # Optional content to be returned after the rendered template.
     #
     # @return [String]
     def output_postamble
-      ""
+      @@default_output_postamble ||= "".html_safe
     end
 
     # Called before rendering the component. Override to perform operations that
@@ -221,6 +237,7 @@ module ViewComponent
     end
 
     if ::Rails.env.development? || ::Rails.env.test?
+      # @private
       def method_missing(method_name, *args) # rubocop:disable Style/MissingRespondToMissing
         super
       rescue => e # rubocop:disable Style/RescueStandardError
@@ -299,6 +316,44 @@ module ViewComponent
 
     def content_evaluated?
       defined?(@__vc_content_evaluated) && @__vc_content_evaluated
+    end
+
+    def maybe_escape_html(text)
+      return text if request && !request.format.html?
+      return text if text.blank?
+
+      if text.html_safe?
+        text
+      else
+        yield
+        html_escape(text)
+      end
+    end
+
+    def safe_render_template_for(variant)
+      if compiler.renders_template_for_variant?(variant)
+        render_template_for(variant)
+      else
+        maybe_escape_html(render_template_for(variant)) do
+          Kernel.warn("WARNING: The #{self.class} component rendered HTML-unsafe output. The output will be automatically escaped, but you may want to investigate.")
+        end
+      end
+    end
+
+    def safe_output_preamble
+      maybe_escape_html(output_preamble) do
+        Kernel.warn("WARNING: The #{self.class} component was provided an HTML-unsafe preamble. The preamble will be automatically escaped, but you may want to investigate.")
+      end
+    end
+
+    def safe_output_postamble
+      maybe_escape_html(output_postamble) do
+        Kernel.warn("WARNING: The #{self.class} component was provided an HTML-unsafe postamble. The postamble will be automatically escaped, but you may want to investigate.")
+      end
+    end
+
+    def compiler
+      @compiler ||= self.class.compiler
     end
 
     # Set the controller used for testing components:
@@ -487,7 +542,9 @@ module ViewComponent
         # Derive the source location of the component Ruby file from the call stack.
         # We need to ignore `inherited` frames here as they indicate that `inherited`
         # has been re-defined by the consuming application, likely in ApplicationComponent.
-        child.source_location = caller_locations(1, 10).reject { |l| l.label == "inherited" }[0].path
+        # We use `base_label` method here instead of `label` to avoid cases where the method
+        # owner is included in a prefix like `ApplicationComponent.inherited`.
+        child.source_location = caller_locations(1, 10).reject { |l| l.base_label == "inherited" }[0].path
 
         # If Rails application is loaded, removes the first part of the path and the extension.
         if defined?(Rails) && Rails.application
@@ -620,7 +677,7 @@ module ViewComponent
 
       # @private
       def collection_counter_parameter
-        "#{collection_parameter}_counter".to_sym
+        :"#{collection_parameter}_counter"
       end
 
       # @private
@@ -630,7 +687,7 @@ module ViewComponent
 
       # @private
       def collection_iteration_parameter
-        "#{collection_parameter}_iteration".to_sym
+        :"#{collection_parameter}_iteration"
       end
 
       # @private