+++ lib/view_component/compiler.rb
@@ -16,6 +16,7 @@ module ViewComponent
     def initialize(component_class)
       @component_class = component_class
       @redefinition_lock = Mutex.new
+      @variants_rendering_templates = Set.new
     end
 
     def compiled?
@@ -56,7 +57,7 @@ module ViewComponent
           RUBY
           # rubocop:enable Style/EvalWithLocation
 
-          component_class.define_method("_call_#{safe_class_name}", component_class.instance_method(:call))
+          component_class.define_method(:"_call_#{safe_class_name}", component_class.instance_method(:call))
 
           component_class.silence_redefinition_of_method("render_template_for")
           component_class.class_eval <<-RUBY, __FILE__, __LINE__ + 1
@@ -68,6 +69,7 @@ module ViewComponent
       else
         templates.each do |template|
           method_name = call_method_name(template[:variant])
+          @variants_rendering_templates << template[:variant]
 
           redefinition_lock.synchronize do
             component_class.silence_redefinition_of_method(method_name)
@@ -84,9 +86,19 @@ module ViewComponent
         define_render_template_for
       end
 
+      component_class.registered_slots.each do |slot_name, config|
+        config[:default_method] = component_class.instance_methods.find { |method_name| method_name == :"default_#{slot_name}" }
+
+        component_class.registered_slots[slot_name] = config
+      end
+
       component_class.build_i18n_backend
 
       CompileCache.register(component_class)
+    end
+
+    def renders_template_for_variant?(variant)
+      @variants_rendering_templates.include?(variant)
     end
 
     private
@@ -101,7 +113,7 @@ module ViewComponent
         "elsif variant.to_sym == :'#{variant}'\n    #{safe_name}"
       end.join("\n")
 
-      component_class.define_method("_call_#{safe_class_name}", component_class.instance_method(:call))
+      component_class.define_method(:"_call_#{safe_class_name}", component_class.instance_method(:call))
 
       body = <<-RUBY
         if variant.nil?
@@ -241,9 +253,9 @@ module ViewComponent
 
     def compiled_inline_template(template)
       handler = ActionView::Template.handler_for_extension(template.language)
-      template.rstrip! if component_class.strip_trailing_whitespace?
+      template = template.source.dup
 
-      compile_template(template.source, handler)
+      compile_template(template, handler)
     end
 
     def compiled_template(file_path)
@@ -258,6 +270,7 @@ module ViewComponent
 
       if handler.method(:call).parameters.length > 1
         handler.call(component_class, template)
+      # :nocov:
       else
         handler.call(
           OpenStruct.new(
@@ -267,6 +280,7 @@ module ViewComponent
           )
         )
       end
+      # :nocov:
     end
 
     def call_method_name(variant)