+++ lib/view_component/engine.rb
@@ -8,8 +8,16 @@ module ViewComponent
   class Engine < Rails::Engine # :nodoc:
     config.view_component = ViewComponent::Config.current
 
-    rake_tasks do
-      load "view_component/rails/tasks/view_component.rake"
+    if Rails.version.to_f < 8.0
+      rake_tasks do
+        load "view_component/rails/tasks/view_component.rake"
+      end
+    else
+      initializer "view_component.stats_directories" do |app|
+        require "rails/code_statistics"
+        dir = ViewComponent::Base.view_component_path
+        Rails::CodeStatistics.register_directory("ViewComponents", dir)
+      end
     end
 
     initializer "view_component.set_configs" do |app|
@@ -80,6 +88,9 @@ module ViewComponent
     initializer "view_component.monkey_patch_render" do |app|
       next if Rails.version.to_f >= 6.1 || !app.config.view_component.render_monkey_patch_enabled
 
+      # :nocov:
+      ViewComponent::Deprecation.deprecation_warning("Monkey patching `render`", "ViewComponent 4.0 will remove the `render` monkey patch")
+
       ActiveSupport.on_load(:action_view) do
         require "view_component/render_monkey_patch"
         ActionView::Base.prepend ViewComponent::RenderMonkeyPatch
@@ -91,10 +102,14 @@ module ViewComponent
         ActionController::Base.prepend ViewComponent::RenderingMonkeyPatch
         ActionController::Base.prepend ViewComponent::RenderToStringMonkeyPatch
       end
+      # :nocov:
     end
 
     initializer "view_component.include_render_component" do |_app|
       next if Rails.version.to_f >= 6.1
+
+      # :nocov:
+      ViewComponent::Deprecation.deprecation_warning("using `render_component`", "ViewComponent 4.0 will remove `render_component`")
 
       ActiveSupport.on_load(:action_view) do
         require "view_component/render_component_helper"
@@ -107,6 +122,7 @@ module ViewComponent
         ActionController::Base.include ViewComponent::RenderingComponentHelper
         ActionController::Base.include ViewComponent::RenderComponentToStringHelper
       end
+      # :nocov:
     end
 
     initializer "static assets" do |app|
@@ -155,6 +171,16 @@ module ViewComponent
           get("_system_test_entrypoint", to: "view_components_system_test#system_test_entrypoint")
         end
       end
+
+      # :nocov:
+      if RUBY_VERSION < "3.0.0"
+        ViewComponent::Deprecation.deprecation_warning("Support for Ruby versions < 3.0.0", "ViewComponent 4.0 will remove support for Ruby versions < 3.0.0 ")
+      end
+
+      if Rails.version.to_f < 6.1
+        ViewComponent::Deprecation.deprecation_warning("Support for Rails versions < 6.1", "ViewComponent 4.0 will remove support for Rails versions < 6.1 ")
+      end
+      # :nocov:
 
       app.executor.to_run :before do
         CompileCache.invalidate! unless ActionView::Base.cache_template_loading