+++ lib/view_component/slotable.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require "active_support/concern"
+require "active_support/inflector/inflections"
 require "view_component/slot"
 
 module ViewComponent
@@ -88,15 +89,15 @@ module ViewComponent
           end
           ruby2_keywords(setter_method_name) if respond_to?(:ruby2_keywords, true)
 
-          define_method slot_name do
+          self::GeneratedSlotMethods.define_method slot_name do
             get_slot(slot_name)
           end
 
-          define_method "#{slot_name}?" do
+          self::GeneratedSlotMethods.define_method :"#{slot_name}?" do
             get_slot(slot_name).present?
           end
 
-          define_method "with_#{slot_name}_content" do |content|
+          define_method :"with_#{slot_name}_content" do |content|
             send(setter_method_name) { content.to_s }
 
             self
@@ -159,7 +160,7 @@ module ViewComponent
           end
           ruby2_keywords(setter_method_name) if respond_to?(:ruby2_keywords, true)
 
-          define_method "with_#{singular_name}_content" do |content|
+          define_method :"with_#{singular_name}_content" do |content|
             send(setter_method_name) { content.to_s }
 
             self
@@ -175,11 +176,11 @@ module ViewComponent
             end
           end
 
-          define_method slot_name do
+          self::GeneratedSlotMethods.define_method slot_name do
             get_slot(slot_name)
           end
 
-          define_method "#{slot_name}?" do
+          self::GeneratedSlotMethods.define_method :"#{slot_name}?" do
             get_slot(slot_name).present?
           end
 
@@ -198,19 +199,28 @@ module ViewComponent
         end
       end
 
-      # Clone slot configuration into child class
-      # see #test_slots_pollution
       def inherited(child)
+        # Clone slot configuration into child class
+        # see #test_slots_pollution
         child.registered_slots = registered_slots.clone
+
+        # Add a module for slot methods, allowing them to be overriden by the component class
+        # see #test_slot_name_can_be_overriden
+        unless child.const_defined?(:GeneratedSlotMethods, false)
+          generated_slot_methods = Module.new
+          child.const_set(:GeneratedSlotMethods, generated_slot_methods)
+          child.include generated_slot_methods
+        end
+
         super
       end
 
       def register_polymorphic_slot(slot_name, types, collection:)
-        define_method(slot_name) do
+        self::GeneratedSlotMethods.define_method(slot_name) do
           get_slot(slot_name)
         end
 
-        define_method("#{slot_name}?") do
+        self::GeneratedSlotMethods.define_method(:"#{slot_name}?") do
           get_slot(slot_name).present?
         end
 
@@ -245,7 +255,7 @@ module ViewComponent
           end
           ruby2_keywords(setter_method_name) if respond_to?(:ruby2_keywords, true)
 
-          define_method "with_#{poly_slot_name}_content" do |content|
+          define_method :"with_#{poly_slot_name}_content" do |content|
             send(setter_method_name) { content.to_s }
 
             self
@@ -265,10 +275,7 @@ module ViewComponent
       end
 
       def define_slot(slot_name, collection:, callable:)
-        # Setup basic slot data
-        slot = {
-          collection: collection
-        }
+        slot = {collection: collection}
         return slot unless callable
 
         # If callable responds to `render_in`, we set it on the slot as a renderable
@@ -295,6 +302,7 @@ module ViewComponent
           raise ReservedPluralSlotNameError.new(name, slot_name)
         end
 
+        raise_if_slot_name_uncountable(slot_name)
         raise_if_slot_conflicts_with_call(slot_name)
         raise_if_slot_ends_with_question_mark(slot_name)
         raise_if_slot_registered(slot_name)
@@ -330,6 +338,13 @@ module ViewComponent
           raise InvalidSlotNameError, "Slot cannot start with 'call_'. Please rename #{slot_name}"
         end
       end
+
+      def raise_if_slot_name_uncountable(slot_name)
+        slot_name = slot_name.to_s
+        if slot_name.pluralize == slot_name.singularize
+          raise UncountableSlotNameError.new(name, slot_name)
+        end
+      end
     end
 
     def get_slot(slot_name)
@@ -357,7 +372,7 @@ module ViewComponent
       # 1. If this is a `content_area` style sub-component, we will render the
       # block via the `slot`
       #
-      # 2. Since we've to pass block content to components when calling
+      # 2. Since we have to pass block content to components when calling
       # `render`, evaluating the block here would require us to call
       # `view_context.capture` twice, which is slower
       slot.__vc_content_block = block if block