+++ test/sandbox/app/components/inline_integer_component.rb
@@ -1,0 +1,5 @@
+class InlineIntegerComponent < ViewComponent::Base
+  def call
+    1
+  end
+end