+++ test/sandbox/app/components/render_check_component.rb
@@ -6,6 +6,6 @@ class RenderCheckComponent < ViewComponent::Base
   end
 
   def call
-    "Rendered"
+    "Rendered".html_safe
   end
 end