+++ test/sandbox/app/components/use_helpers_component.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
 class UseHelpersComponent < ViewComponent::Base
-  include ViewComponent::UseHelpers
-
-  use_helpers :message
+  use_helper :message
 end