+++ test/sandbox/config/routes.rb
@@ -29,10 +29,14 @@ Sandbox::Application.routes.draw do
   get :cached_partial, to: "integration_examples#cached_partial"
   get :inherited_sidecar, to: "integration_examples#inherited_sidecar"
   get :inherited_from_uncompilable_component, to: "integration_examples#inherited_from_uncompilable_component"
+  get :unsafe_component, to: "integration_examples#unsafe_component"
+  get :unsafe_preamble_component, to: "integration_examples#unsafe_preamble_component"
+  get :unsafe_postamble_component, to: "integration_examples#unsafe_postamble_component"
   post :create, to: "integration_examples#create"
 
   constraints(lambda { |request| request.env["warden"].authenticate! }) do
     get :constraints_with_env, to: "integration_examples#index"
   end
   get :helpers_proxy_component, to: "integration_examples#helpers_proxy_component"
+  get :turbo_stream, to: "integration_examples#turbo_stream"
 end