+++ test/test_engine/lib/test_engine/engine.rb
@@ -1,0 +1,5 @@
+module TestEngine
+  class Engine < ::Rails::Engine
+    isolate_namespace TestEngine
+  end
+end